feat(S7b): resample_poly port + f64 resample fix — bit-exact non-250 parity (ADR 0069)
Second S7b increment: port `resample_poly` (the common resample→250 branch) to
Rust, and fix a float32 precision artifact so the whole non-250 pipeline is
bit-exact to Python.

Rust (`src/normalize.rs`):
- `resample_poly` = firwin Kaiser-windowed sinc (`i0` Bessel series + `sinc` +
  DC-normalize, validated to 5.5e-17 vs scipy.firwin) → `upfirdn` polyphase →
  scipy's center-trim padding. `resample_to_250` branches on the gcd-reduced
  up/down (poly when both ≤256, else `NormalizeError::FftResampleUnsupported` —
  the FFT branch is a later increment). `normalize_eeg(data, orig_sr)` chains
  resample→250 → 0.5 Hz zero-phase HP → Q31. Unit tests validate against scipy
  oracles (200→250, 500→250) + branch logic.

f64 resample fix (user architectural decision — "fix production to f64 too"):
- `lma_dataset.py` previously fed float32 into `resample_poly` (channel-select
  preserves the f32 calibration) and only upcast to f64 AFTER — so scipy
  resampled in float32, a ~2e-7-relative artifact (below the ADC noise floor,
  but real). Cast to f64 UP FRONT now, in both `lma_dataset.py` and the golden
  dumper, matching the Rust f64 port. This is a one-time, below-noise-floor
  change to non-250 Hz training inputs; the 250 Hz path is untouched (no
  resample).

Parity gate: added the 200 Hz (poly-branch) case to `normalize_parity.rs` +
`dump_normalize_golden.py`. With both sides in f64, BOTH cases are now
**BIT-EXACT (max|Δ|=0)** — 250 Hz (HP+Q31) and 200 Hz (resample+HP+Q31). The
200 Hz tolerance is 4 LSB of headroom for cross-libm `sin` ULP drift in the
transcendental filter design; it reports the actual delta (0) for monitoring.

Remaining S7b: FFT resample branch (rare odd rates), channel_resolver 10-20
alias port, the `normalize_eeg` pyfunction, and the lma_dataset.py cutover.
core-no_std untouched (host-only module); frozen wire gates unaffected.

+++ reference_implementations/python_codec/lamquant_codec/training/lma_dataset.py
@@ -406,6 +406,13 @@ def decode_lma_signal(lma_path: str, stem: str,
         data = padded
 
     if abs(original_sr - TARGET_SR) > 0.5:
+        # ADR 0069 S7b: resample in float64, not float32. Previously `data` was
+        # float32 here (channel-select preserves the f32 calibration), so
+        # `resample_poly`/`resample` ran in float32 and the `.astype(float64)`
+        # only upcast AFTER the precision loss. Casting up FRONT makes the Rust
+        # f64 port (src/normalize.rs) bit-exact to this path and removes a
+        # ~2e-7-relative artifact (below the ADC noise floor, but free to fix).
+        data = data.astype(np.float64)
         up = int(TARGET_SR)
         down = int(original_sr)
         g = gcd(up, down)
@@ -418,7 +425,7 @@ def decode_lma_signal(lma_path: str, stem: str,
                 resampled[ch] = resample(data[ch], new_len)
             data = resampled
         else:
-            data = _LAZY["resample_poly"](data, up, down, axis=1).astype(np.float64)
+            data = _LAZY["resample_poly"](data, up, down, axis=1)
 
     data = sosfiltfilt(_highpass_sos(), data, axis=1)
